+++ web/src/main/java/com/weicoder/web/socket/interfaces/Client.java
@@ -17,7 +17,7 @@ public interface Client extends Socket {
 	 * @param id 指令
 	 * @param message 消息
 	 */
-	void send(int id, Object message);
+	void send(short id, Object message);
 
 	/**
 	 * 写入数据
